+++ services/madoc-ts/src/frontend/shared/capture-models/hooks/use-model-translation.tsx
@@ -3,6 +3,8 @@ import { useTranslation, UseTranslationOptions } from 'react-i18next';
 
 const ModelNamespace = createContext('capture-models');
 
+ModelNamespace.displayName = 'ModelNamespace';
+
 export function WithModelNamespace(props: { namespace: string; children: ReactElement }) {
   return <ModelNamespace.Provider value={props.namespace}>{props.children}</ModelNamespace.Provider>;
 }