+++ services/madoc-ts/src/frontend/shared/hooks/use-site.ts
@@ -24,6 +24,8 @@ const SiteReactContext = React.createContext<
   | undefined
 >(undefined);
 
+SiteReactContext.displayName = 'Site';
+
 export const useSiteTheme = () => {
   const details = useContext(SiteReactContext);